% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cree_repertoires.R
\name{cree_repertoires}
\alias{cree_repertoires}
\title{on cree les repertoires 2_data, 4_resultats dans le nouveau dossier s ils n'existent pas}
\usage{
cree_repertoires(dossier = here::here())
}
\arguments{
\item{dossier}{le repertoire où est le projet}
}
\value{
cree les sous-repertoires
}
\description{
on cree les repertoires 2_data, 4_resultats dans le nouveau dossier s ils n'existent pas
}
