% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn01_import_tabgeo.R
\name{fn01_import_tabgeo}
\alias{fn01_import_tabgeo}
\title{fn01_import_tabgeo}
\usage{
fn01_import_tabgeo(x = "table_supracom_2020")
}
\arguments{
\item{x}{nom de la table}
}
\value{
tibble tab_geo
}
\description{
fn01_import_tabgeo
}
\examples{
# table des epci 2020
fn01_import_tabgeo(x = "table_supracom_2020")
}
