% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn02_import_tab_dido.R
\name{fn02_import_tab_dido}
\alias{fn02_import_tab_dido}
\title{fn02_import_tab_dido}
\usage{
fn02_import_tab_dido(x = "ommercialisation")
}
\arguments{
\item{x}{ommercialisation par defaut}
}
\value{
nothing
}
\description{
fn02_import_tab_dido
}
