% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn02_import_tab_rd1.R
\name{fn02_import_tab_rd1}
\alias{fn02_import_tab_rd1}
\title{fn02_import_tab_rd1}
\usage{
fn02_import_tab_rd1(x = "2_data")
}
\arguments{
\item{x}{le nom du repertoire ou est le fichier}
}
\value{
nothing
}
\description{
fn02_import_tab_rd1
}
