% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn03_transforme_requete.R
\name{fn03_transforme_requete}
\alias{fn03_transforme_requete}
\title{fn03_transforme_requete}
\usage{
fn03_transforme_requete(x = "Indicateurs_ecln_trim2022.xlsx")
}
\arguments{
\item{x}{le nom du fichier issu de geokit}
}
\value{
nothing
}
\description{
fn03_transforme_requete
}
