% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn04_verifie_communes_manquantes.R
\name{fn04_verifie_communes_manquantes}
\alias{fn04_verifie_communes_manquantes}
\title{fn04_verifie_communes_manquantes}
\usage{
fn04_verifie_communes_manquantes(data = ls_onglets)
}
\arguments{
\item{data}{df}
}
\value{
liste
}
\description{
fn04_verifie_communes_manquantes
}
