% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn05_complete_table_communes.R
\name{fn05_complete_table_communes}
\alias{fn05_complete_table_communes}
\title{fn05_complete_table_communes}
\usage{
fn05_complete_table_communes(data = ls_onglets$cor_com)
}
\arguments{
\item{data}{tibble}
}
\value{
df
}
\description{
fn05_complete_table_communes
}
