% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn99_utilitaire_importFrom.R
\name{fn99_utilitaire_importFrom}
\alias{fn99_utilitaire_importFrom}
\title{utilitaire_importFrom}
\usage{
fn99_utilitaire_importFrom(x = "fn06_met_en_forme_liste.R")
}
\arguments{
\item{x}{nom d'un fichier de fonction}
}
\value{
les packages et fonctions utilisées
}
\description{
utile pour lister les fonctions externes utilisees par chaque fonction
}
