% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liste_modifie_date.R
\name{liste_modifie_date}
\alias{liste_modifie_date}
\title{A partir des onglets du tableau excel copies dans une liste de tableaux
on va transformer d'un seul coup la date dans tous les tableaux}
\usage{
liste_modifie_date(data)
}
\arguments{
\item{data}{une liste de df avec chacune une colonne date au format 20201}
}
\description{
A partir des onglets du tableau excel copies dans une liste de tableaux
on va transformer d'un seul coup la date dans tous les tableaux
}
