% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforme_tab_rd1.R
\name{transforme_tab_rd1}
\alias{transforme_tab_rd1}
\title{transforme_tab_rd1}
\usage{
transforme_tab_rd1(x = "_rd1_commercialisation")
}
\arguments{
\item{x}{un vecteur de characteres}
}
\value{
le tableau mis en forme copie dans le repetoire 4_resultats
}
\description{
fonction qui va transformer le tableau rd1_commercialisation
}
