% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforme_tableaux.R
\name{transforme_tableaux}
\alias{transforme_tableaux}
\title{Fonction principale va transformer le tableau excel en plusieurs csv
les fichiers csv vont etre enregistres dans 4_resultats}
\usage{
transforme_tableaux(x = "Indicateurs_ecln_trim.xlsx")
}
\arguments{
\item{x}{le nom du fichier excel à modifier}
}
\value{
plusieurs fichiers csv mis en forme sont copiés dans 4_resultats
}
\description{
Fonction principale va transformer le tableau excel en plusieurs csv
les fichiers csv vont etre enregistres dans 4_resultats
}
