% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifie_geo_manquant.R
\name{verifie_geo_manquant}
\alias{verifie_geo_manquant}
\title{Verifie les territoires manquants}
\usage{
verifie_geo_manquant(data)
}
\arguments{
\item{data}{a list}
}
\value{
a list
}
\description{
Verifie les territoires manquants
}
